refactor: split integration tests into multiple files
